+++ cpp/arcticdb/processing/expression_node.cpp
@@ -7,7 +7,9 @@
 
 #include <arcticdb/processing/expression_node.hpp>
 #include <arcticdb/processing/processing_segment.hpp>
-#include <arcticdb/processing/operation_dispatch.hpp>
+#include <arcticdb/processing/operation_types.hpp>
+#include <arcticdb/processing/operation_dispatch_binary.hpp>
+#include <arcticdb/processing/operation_dispatch_unary.hpp>
 
 namespace arcticdb {
